[x86_64] Port SCU IPC and Watchdog driver
BZ: 158551

In order to support the 64 bits platform, we fix some wrong type
usages and some bad practices that are in used in the SCU IPC and
watchdog driver. Here is highlight of what this patch does:
 - use the phys_addr_t type when appropriate (instead of u32 or
   unsigned int or even long ...) ;
 - use "void" type pointer instead of "u32" type pointers ;
 - use the dedicated "%pa" printk format to print phys_addr_t
   variables

Not directly related to 64 bits, but had to be cleaned too:
 - repectively replace the structure member types __u32, __u16 and
   __u8 with the u32, u16, u8 types in structures which does not
   interract with user-space ;
 - move some variables from the heap to the stack by removing the
   static keyword when it is inappropriate ;

+++ drivers/staging/android/alarm-dev.c
@@ -259,8 +259,7 @@ static long alarm_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 {
 
 	struct timespec ts;
-	/* Using u32 instead of int for 32bit uspace and 64bit kernel */
-	u32 rv = 0;
+	long rv = 0;
 
 
 	switch (ANDROID_ALARM_BASE_CMD(cmd)) {
@@ -291,7 +290,7 @@ static long alarm_compat_ioctl(struct file *file, unsigned int cmd,
 {
 
 	struct timespec ts;
-	int rv;
+	long rv;
 
 	switch (ANDROID_ALARM_BASE_CMD(cmd)) {
 	case ANDROID_ALARM_SET_AND_WAIT_COMPAT(0):
